+++ src/main/java/com/cricket/local_score/Repository/PlayerRepository.java
@@ -8,7 +8,7 @@ import org.springframework.stereotype.Repository;
 import java.util.List;
 
 @Repository
-public interface PlayerRepository extends JpaRepository<PlayerEntity,Integer> {
+public interface PlayerRepository extends JpaRepository<PlayerEntity,Long> {
     List<PlayerEntity> findByName(String name);
     // List<PlayerEntity> findByNameContainingIgnoreCase(String name);
 }
